fix bugs

+++ backend-t-t/app/Application/Services/Tasks/TaskService.php
@@ -1,4 +1,5 @@
-<?php 
+<?php
+
 namespace App\Application\Services\Tasks;
 
 use App\Domain\Interfaces\MemberRepositoryInterface;
@@ -6,28 +7,32 @@ use App\Domain\Interfaces\TaskRepositoryInterface;
 use App\Domain\Interfaces\TeamRepositoryInterface;
 use App\Models\TeamMember;
 
-class TaskService{
+class TaskService
+{
 
-    public function __construct(private TaskRepositoryInterface $taskRepositoryInterface,
-    private MemberRepositoryInterface $memberRepositoryInterface){}
+    public function __construct(
+        private TaskRepositoryInterface $taskRepositoryInterface,
+        private MemberRepositoryInterface $memberRepositoryInterface
+    ) {}
 
-    public function create($request){
+    public function create($request, $team)
+    {
 
         $validatedData = $request->validated();
 
-        $teamMember = $this->memberRepositoryInterface->findTeamMember($validatedData['team_id'],$request->user()->id);
+        $teamMember = $this->memberRepositoryInterface->findTeamMember($team->id, $request->user()->id);
 
-        if(!$teamMember){
+        if (!$teamMember) {
             throw new \Exception("You are not a member of this team");
-            
         }
         $validatedData['team_member_id'] = $teamMember->id;
+        $validatedData['team_id'] = $team->id;
+
 
         $task = $this->taskRepositoryInterface->create($validatedData);
 
-        if(!$task){
+        if (!$task) {
             throw new \Exception("Task is not created");
-            
         }
 
         return [
@@ -36,17 +41,25 @@ class TaskService{
             'message' => 'Task created successfully',
             'data' => $task
         ];
-
-
     }
 
-    public function update($request, $task){
+    public function update($request, $task, $team)
+    {
         
+        if (!$team || !$task) {
+            throw new \Exception("Team or task not found");
+        }
+        if (!$team->teamMembers()->where('user_id', $request->user()->id)->exists()) {
+            throw new \Exception("You are not a member of this team");
+        }
+        if (!$task->team_id == $team->id) {
+            throw new \Exception("Task is not associated with this team");
+        }
+
         $task = $this->taskRepositoryInterface->update($task, $request->validated());
 
-        if(!$task){
+        if (!$task) {
             throw new \Exception("Task update is not successful");
-            
         }
 
         return [
@@ -57,8 +70,18 @@ class TaskService{
         ];
     }
 
-    public function delete($task){
-        
+    public function delete($team, $task)
+    {
+        if(!$team || !$task){
+            throw new \Exception("Team or task not found");
+        }
+        if(!$team->teamMembers()->where('user_id', auth()->user()->id)->exists()){
+            throw new \Exception("You are not a member of this team");
+        }
+        if(!$task->team_id == $team->id){
+            throw new \Exception("Task is not associated with this team");
+        }
+
         $this->taskRepositoryInterface->delete($task);
 
         return [
@@ -68,12 +91,22 @@ class TaskService{
         ];
     }
 
-    public function tasks($request){
-        
+    public function tasks($request, $team)
+    {
 
-         $tasks = $this->taskRepositoryInterface->tasks([
+
+        if(!$team){
+            throw new \Exception("Team not found");
+        }
+
+        if(!$team->teamMembers()->where('user_id', $request->user()->id)->exists()){
+            throw new \Exception("You are not a member of this team");
+        }
+
+
+        $tasks = $this->taskRepositoryInterface->tasks([
             'user_id'    => $request->input('user_id'),
-            'team_id'    => $request->input('team_id'),
+            'team_id'    => $team->id,
             'priority'   => $request->input('priority'),
             'status'     => $request->input('status'),
             'category'   => $request->input('category'),
@@ -83,7 +116,7 @@ class TaskService{
 
         $taskLists = [];
 
-        foreach($tasks as $task){
+        foreach ($tasks as $task) {
             $taskLists[] = [
                 'id' => $task->id,
                 'title' => $task->title,
@@ -105,7 +138,8 @@ class TaskService{
         ];
     }
 
-    public function list($teamId){
+    public function list($teamId)
+    {
         $tasks = $this->taskRepositoryInterface->listByTeam($teamId);
 
         return [
@@ -116,7 +150,8 @@ class TaskService{
         ];
     }
 
-    public function selfTasks(){
+    public function selfTasks()
+    {
         $tasks = $this->taskRepositoryInterface->listByMember(auth()->user()->id);
 
         return [
@@ -127,17 +162,18 @@ class TaskService{
         ];
     }
 
-    public function view($team, $task){
+    public function view($team, $task)
+    {
 
-        if(!$team || !$task){
+        if (!$team || !$task) {
             throw new \Exception("Team or task not found");
         }
 
-        if(!$team->teamMembers()->where('user_id', auth()->user()->id)->exists()){
+        if (!$team->teamMembers()->where('user_id', auth()->user()->id)->exists()) {
             throw new \Exception("You are not a member of this team");
         }
-        
-        if(!$task->team_id == $team->id){
+
+        if (!$task->team_id == $team->id) {
             throw new \Exception("Task is not associated with this team");
         }
